Time disappearing for momentjs
This is a fix for the time disappearing on x-axis when momentjs and moment-timezone are used. This happened because Dygraphs uses hours in the form 0-23 for current day, then 24,25,26... for the next day. Parsing the hour in a such a format led to a breaking in momentjs. What I have done is simply change it to 0-23 again and handle the corresponding changes for day, month and year changes.